Removing old PBF ingest logic, updating unit and integration tests
