+++ src/integrationTest/java/org/openstreetmap/atlas/geography/atlas/multi/MultiAtlasIntegrationTest.java
@@ -111,7 +111,7 @@ public class MultiAtlasIntegrationTest extends AtlasIntegrationTest
 
         final Polygon polygon = this.multi.area(24601488000000L).asPolygon();
 
-        Assert.assertEquals(6, Iterables.size(this.multi.edgesIntersecting(polygon)));
+        Assert.assertEquals(40, Iterables.size(this.multi.edgesIntersecting(polygon)));
         Assert.assertEquals(7, Iterables.size(this.multi.areasIntersecting(polygon)));
     }
 }
