+++ src/integrationTest/java/org/openstreetmap/atlas/geography/atlas/pbf/slicing/WaySectionProcessorIntegrationTest.java
@@ -8,6 +8,7 @@ import org.openstreetmap.atlas.geography.Rectangle;
 import org.openstreetmap.atlas.geography.atlas.Atlas;
 import org.openstreetmap.atlas.geography.atlas.AtlasIntegrationTest;
 import org.openstreetmap.atlas.geography.atlas.items.Edge;
+import org.openstreetmap.atlas.geography.atlas.items.Line;
 import org.openstreetmap.atlas.geography.atlas.pbf.AtlasLoadingOption;
 import org.openstreetmap.atlas.geography.atlas.pbf.slicing.identifier.AbstractIdentifierFactory;
 import org.openstreetmap.atlas.geography.atlas.pbf.slicing.identifier.PaddingIdentifierFactory;
@@ -19,8 +20,8 @@ import org.openstreetmap.atlas.utilities.collections.Iterables;
  */
 public class WaySectionProcessorIntegrationTest extends AtlasIntegrationTest
 {
-    private static Atlas unSlicedAtlas;
-    private static Atlas slicedAtlas;
+    private static Atlas rawAtlas;
+    private static Atlas sectionedAtlas;
 
     @Before
     public void createAtlas()
@@ -28,53 +29,59 @@ public class WaySectionProcessorIntegrationTest extends AtlasIntegrationTest
         final Rectangle belizeCity = Rectangle.forLocated(
                 Location.forString("17.521983, -88.213739"),
                 Location.forString("17.491327, -88.178071"));
-        if (unSlicedAtlas == null)
+
+        // Both atlases are not country-sliced
+        if (rawAtlas == null)
         {
-            unSlicedAtlas = loadBelizeRaw(belizeCity, AtlasLoadingOption.createOptionWithNoSlicing()
+            rawAtlas = loadBelizeRaw(belizeCity, AtlasLoadingOption.createOptionWithNoSlicing()
                     .setLoadWaysSpanningCountryBoundaries(false));
         }
-        if (slicedAtlas == null)
+        if (sectionedAtlas == null)
         {
-            slicedAtlas = loadBelizeRaw(belizeCity, AtlasLoadingOption.createOptionWithNoSlicing()
-                    .setWaySectioning(true).setLoadWaysSpanningCountryBoundaries(false));
+            sectionedAtlas = loadBelizeRaw(belizeCity,
+                    AtlasLoadingOption.createOptionWithNoSlicing().setWaySectioning(true)
+                            .setLoadWaysSpanningCountryBoundaries(false));
         }
     }
 
     @Test
     public void testPrimaryWay()
     {
-        // Primary way should be sliced
-        final long primaryWayShouldBeSliced = 279042065L;
+        // Primary way should be sectioned
+        final long primaryWayShouldBeSectioned = 279042065L;
         final AbstractIdentifierFactory factory = new WaySectionIdentifierFactory(
-                PaddingIdentifierFactory.pad(primaryWayShouldBeSliced));
+                PaddingIdentifierFactory.pad(primaryWayShouldBeSectioned));
         Assert.assertEquals(21,
-                Iterables.size(unSlicedAtlas.edge(primaryWayShouldBeSliced).getRawGeometry()));
-        Assert.assertNull(slicedAtlas.edge(factory.getReferenceIdentifier()));
+                Iterables.size(rawAtlas.line(primaryWayShouldBeSectioned).getRawGeometry()));
+        Assert.assertNull(sectionedAtlas.edge(factory.getReferenceIdentifier()));
         Assert.assertEquals(2,
-                Iterables.size(slicedAtlas.edge(factory.nextIdentifier()).getRawGeometry()));
+                Iterables.size(sectionedAtlas.edge(factory.nextIdentifier()).getRawGeometry()));
         Assert.assertEquals(2,
-                Iterables.size(slicedAtlas.edge(factory.nextIdentifier()).getRawGeometry()));
+                Iterables.size(sectionedAtlas.edge(factory.nextIdentifier()).getRawGeometry()));
         Assert.assertEquals(2,
-                Iterables.size(slicedAtlas.edge(factory.nextIdentifier()).getRawGeometry()));
-        Assert.assertEquals(2, Iterables.size(slicedAtlas.edge(279042065000015L).getRawGeometry()));
-        Assert.assertEquals(5, Iterables.size(slicedAtlas.edge(279042065000016L).getRawGeometry()));
-        Assert.assertEquals(2, Iterables.size(slicedAtlas.edge(279042065000017L).getRawGeometry()));
-        Assert.assertNull(slicedAtlas.edge(279042065000018L));
+                Iterables.size(sectionedAtlas.edge(factory.nextIdentifier()).getRawGeometry()));
+        Assert.assertEquals(2,
+                Iterables.size(sectionedAtlas.edge(279042065000015L).getRawGeometry()));
+        Assert.assertEquals(5,
+                Iterables.size(sectionedAtlas.edge(279042065000016L).getRawGeometry()));
+        Assert.assertEquals(2,
+                Iterables.size(sectionedAtlas.edge(279042065000017L).getRawGeometry()));
+        Assert.assertNull(sectionedAtlas.edge(279042065000018L));
     }
 
     @Test
     public void testSelfIntersectionRing()
     {
-        // Way 295734091 will be sliced into 295734091000001 and 295734091000002. However because
-        // 295734091000002 is a ring, which should be further sliced into 295734091000003 and
+        // Way 295734091 will be sectioned into 295734091000001 and 295734091000002. However because
+        // 295734091000002 is a ring, which should be further sectioned into 295734091000003 and
         // 295734091000004
         final long selfIntersectionWay = 295734091;
-        Assert.assertNotNull(unSlicedAtlas.edge(selfIntersectionWay));
+        Assert.assertNotNull(rawAtlas.line(selfIntersectionWay));
         final AbstractIdentifierFactory factory = new WaySectionIdentifierFactory(
                 PaddingIdentifierFactory.pad(selfIntersectionWay));
-        Assert.assertNull(slicedAtlas.edge(factory.getReferenceIdentifier()));
-        Assert.assertNotNull(slicedAtlas.edge(factory.nextIdentifier()));
-        Assert.assertNotNull(slicedAtlas.edge(factory.nextIdentifier()));
+        Assert.assertNull(sectionedAtlas.edge(factory.getReferenceIdentifier()));
+        Assert.assertNotNull(sectionedAtlas.edge(factory.nextIdentifier()));
+        Assert.assertNotNull(sectionedAtlas.edge(factory.nextIdentifier()));
     }
 
     @Test
@@ -83,58 +90,59 @@ public class WaySectionProcessorIntegrationTest extends AtlasIntegrationTest
         // The identifier order of split way should be same as the order of shape point for way with
         // tag oneway=-1
         final long reversedWay = 25977940L;
-        final Edge nonSplit = unSlicedAtlas.edge(reversedWay);
+        final Line nonSplit = rawAtlas.line(reversedWay);
         Assert.assertNotNull(nonSplit);
         final AbstractIdentifierFactory factory = new WaySectionIdentifierFactory(
                 PaddingIdentifierFactory.pad(reversedWay));
-        Assert.assertNull(slicedAtlas.edge(factory.getReferenceIdentifier()));
-        final Edge firstSplit = slicedAtlas.edge(factory.nextIdentifier());
-        final Edge secondSplit = slicedAtlas.edge(factory.nextIdentifier());
+        Assert.assertNull(sectionedAtlas.edge(factory.getReferenceIdentifier()));
+
+        final Edge firstSplit = sectionedAtlas.edge(factory.nextIdentifier());
+        final Edge secondSplit = sectionedAtlas.edge(factory.nextIdentifier());
+
         Assert.assertNotNull(firstSplit);
         Assert.assertNotNull(secondSplit);
-        Assert.assertEquals(nonSplit.start().getLocation(), firstSplit.start().getLocation());
-        Assert.assertEquals(nonSplit.end().getLocation(), secondSplit.end().getLocation());
+        Assert.assertEquals(nonSplit.asPolyLine().last(), secondSplit.start().getLocation());
+        Assert.assertEquals(nonSplit.asPolyLine().first(), firstSplit.end().getLocation());
     }
 
     @Test
-    public void testTotalNumber()
+    public void testTotalCounts()
     {
-        // Total Numbers
-        Assert.assertEquals(1020, unSlicedAtlas.numberOfEdges());
-        Assert.assertEquals(2843, slicedAtlas.numberOfEdges());
+        Assert.assertEquals(0, rawAtlas.numberOfEdges());
+        Assert.assertEquals(2841, sectionedAtlas.numberOfEdges());
 
-        Assert.assertEquals(8, unSlicedAtlas.numberOfLines());
-        Assert.assertEquals(8, slicedAtlas.numberOfLines());
+        Assert.assertEquals(917, rawAtlas.numberOfLines());
+        Assert.assertEquals(8, sectionedAtlas.numberOfLines());
 
-        Assert.assertEquals(959, unSlicedAtlas.numberOfNodes());
-        Assert.assertEquals(1167, slicedAtlas.numberOfNodes());
+        Assert.assertEquals(0, rawAtlas.numberOfNodes());
+        Assert.assertEquals(1165, sectionedAtlas.numberOfNodes());
 
-        Assert.assertEquals(336, unSlicedAtlas.numberOfAreas());
-        Assert.assertEquals(336, slicedAtlas.numberOfAreas());
+        Assert.assertEquals(0, rawAtlas.numberOfAreas());
+        Assert.assertEquals(336, sectionedAtlas.numberOfAreas());
 
-        Assert.assertEquals(1, unSlicedAtlas.numberOfRelations());
-        Assert.assertEquals(1, slicedAtlas.numberOfRelations());
+        Assert.assertEquals(1, rawAtlas.numberOfRelations());
+        Assert.assertEquals(1, sectionedAtlas.numberOfRelations());
 
-        Assert.assertEquals(82, unSlicedAtlas.numberOfPoints());
-        Assert.assertEquals(82, slicedAtlas.numberOfPoints());
+        Assert.assertEquals(5165, rawAtlas.numberOfPoints());
+        Assert.assertEquals(82, sectionedAtlas.numberOfPoints());
     }
 
     @Test
     public void testWaterWay()
     {
-        // Waterway (Atlas area) should not be sliced
+        // Waterway (Atlas area) should not be sectioned
         final long waterway = 25977495L;
-        Assert.assertNotNull(unSlicedAtlas.area(waterway));
-        Assert.assertNotNull(slicedAtlas.area(PaddingIdentifierFactory.pad(waterway)));
+        Assert.assertNotNull(rawAtlas.line(waterway));
+        Assert.assertNotNull(sectionedAtlas.area(PaddingIdentifierFactory.pad(waterway)));
     }
 
     @Test
     public void testWayAcrossArea()
     {
-        // Way across an area should not be sliced
+        // Way across an area should not be sectioned
         final long wayAcrossArea = 194237644L;
-        Assert.assertEquals(2, Iterables.size(unSlicedAtlas.edge(wayAcrossArea).getRawGeometry()));
+        Assert.assertEquals(2, Iterables.size(rawAtlas.line(wayAcrossArea).getRawGeometry()));
         Assert.assertEquals(2, Iterables.size(
-                slicedAtlas.edge(PaddingIdentifierFactory.pad(wayAcrossArea)).getRawGeometry()));
+                sectionedAtlas.edge(PaddingIdentifierFactory.pad(wayAcrossArea)).getRawGeometry()));
     }
 }
