+++ src/main/java/org/openstreetmap/atlas/geography/atlas/converters/AtlasDebugTool.java
@@ -9,7 +9,9 @@ import org.openstreetmap.atlas.geography.Rectangle;
 import org.openstreetmap.atlas.geography.atlas.Atlas;
 import org.openstreetmap.atlas.geography.atlas.AtlasResourceLoader;
 import org.openstreetmap.atlas.geography.atlas.pbf.AtlasLoadingOption;
-import org.openstreetmap.atlas.geography.atlas.pbf.OsmPbfLoader;
+import org.openstreetmap.atlas.geography.atlas.raw.creation.RawAtlasGenerator;
+import org.openstreetmap.atlas.geography.atlas.raw.sectioning.WaySectionProcessor;
+import org.openstreetmap.atlas.geography.atlas.raw.slicing.RawAtlasCountrySlicer;
 import org.openstreetmap.atlas.geography.atlas.routing.AStarRouter;
 import org.openstreetmap.atlas.geography.boundary.CountryBoundaryMap;
 import org.openstreetmap.atlas.geography.boundary.converters.CountryListTwoWayStringConverter;
@@ -77,7 +79,7 @@ public class AtlasDebugTool extends Command
         @SuppressWarnings("unchecked")
         final List<Location> startEndRoute = (List<Location>) command.get(ROUTE);
 
-        final Atlas atlas;
+        Atlas atlas;
         if (pbf != null && pbf.exists())
         {
             final AtlasLoadingOption option;
@@ -108,7 +110,14 @@ public class AtlasDebugTool extends Command
             {
                 multiPolygon = inputMultipolygon;
             }
-            atlas = new OsmPbfLoader(pbf, multiPolygon, option).read();
+
+            atlas = new RawAtlasGenerator(pbf, option, multiPolygon).build();
+            if (option.isCountrySlicing())
+            {
+                atlas = new RawAtlasCountrySlicer(option.getCountryCodes(),
+                        option.getCountryBoundaryMap()).slice(atlas);
+            }
+            atlas = new WaySectionProcessor(atlas, option).run();
             atlas.save(atlasFile);
         }
         else if (atlasFile != null && atlasFile.exists())
