+++ src/main/java/org/openstreetmap/atlas/geography/boundary/CountryBoundaryMap.java
@@ -40,8 +40,8 @@ import org.openstreetmap.atlas.geography.atlas.Atlas;
 import org.openstreetmap.atlas.geography.atlas.items.AtlasEntity;
 import org.openstreetmap.atlas.geography.atlas.items.complex.boundaries.ComplexBoundary;
 import org.openstreetmap.atlas.geography.atlas.items.complex.boundaries.ComplexBoundaryFinder;
-import org.openstreetmap.atlas.geography.atlas.pbf.slicing.RuntimeCounter;
 import org.openstreetmap.atlas.geography.atlas.raw.slicing.CountryCodeProperties;
+import org.openstreetmap.atlas.geography.atlas.raw.slicing.RuntimeCounter;
 import org.openstreetmap.atlas.geography.boundary.converters.CountryListTwoWayStringConverter;
 import org.openstreetmap.atlas.geography.converters.jts.JtsMultiPolygonConverter;
 import org.openstreetmap.atlas.geography.converters.jts.JtsMultiPolygonToMultiPolygonConverter;
